refactor: Move camera changed demo widgets to widgets package
- Moved WhiskeyCompass.kt to widgets package.
- Extracted TiltScale and RangeScale to separate files in widgets package.
- Updated CameraChangedActivity to use widgets from the new package.
