+++ maps3d-compose-demo/src/main/java/com/example/maps3dcomposedemo/CameraChangedActivity.kt
@@ -48,6 +48,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.google.android.gms.maps3d.model.Camera
 import com.google.android.gms.maps3d.model.Map3DMode
+import com.example.maps3dcomposedemo.widgets.WhiskeyCompass
+import com.example.maps3dcomposedemo.widgets.TiltScale
+import com.example.maps3dcomposedemo.widgets.RangeScale
 import com.google.android.gms.maps3d.model.camera
 import com.google.android.gms.maps3d.model.latLngAltitude
 import com.google.maps.android.compose3d.GoogleMap3D
@@ -134,77 +137,4 @@ fun CameraChangedScreen() {
     }
 }
 
-@Composable
-fun TiltScale(tilt: Float, modifier: Modifier = Modifier) {
-    // User requested: 90 = straight down, 0 = horizontal
-    // SDK: 0 = straight down, 90 = horizontal
-    // So displayedTilt = 90 - sdkTilt
-    val displayedTilt = 90f - tilt
-    val textMeasurer = rememberTextMeasurer()
-    val onPrimaryColor = MaterialTheme.colorScheme.onPrimaryContainer
 
-    Box(
-        modifier = modifier
-            .height(300.dp)
-            .width(80.dp)
-            .background(MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.8f))
-            .clipToBounds(),
-        contentAlignment = Alignment.Center
-    ) {
-        Canvas(modifier = Modifier.fillMaxSize()) {
-            val centerLineY = size.height / 2f
-            val centerLineX = size.width / 2f
-            
-            // Pixels per degree
-            val pixelsPerDegree = 5f
-            
-            // Draw a fixed center line (pointer)
-            drawLine(
-                color = Color.Red,
-                start = Offset(0f, centerLineY),
-                end = Offset(size.width, centerLineY),
-                strokeWidth = 2f
-            )
-
-            // Draw scrolling ticks
-            translate(top = centerLineY + (displayedTilt * pixelsPerDegree)) {
-                for (i in 0..90 step 5) {
-                    val yPos = -i * pixelsPerDegree
-                    
-                    val isMajor = i % 15 == 0
-                    val tickLength = if (isMajor) 15f else 8f
-                    
-                    drawLine(
-                        color = onPrimaryColor,
-                        start = Offset(centerLineX - tickLength, yPos),
-                        end = Offset(centerLineX + tickLength, yPos),
-                        strokeWidth = if (isMajor) 2f else 1f
-                    )
-                    
-                    if (isMajor) {
-                        val measuredText = textMeasurer.measure(i.toString(), style = TextStyle(color = onPrimaryColor, fontSize = 12.sp))
-                        drawText(
-                            textLayoutResult = measuredText,
-                            topLeft = Offset(centerLineX + 20f, yPos - measuredText.size.height / 2f)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun RangeScale(range: Float, modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .background(MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.8f))
-            .padding(8.dp)
-    ) {
-        Text(
-            text = "Range: ${range.roundToInt()} m",
-            color = MaterialTheme.colorScheme.onPrimaryContainer,
-            style = MaterialTheme.typography.bodyMedium
-        )
-    }
-}
