+++ maps3d-compose-demo/src/main/java/com/example/maps3dcomposedemo/widgets/RangeScale.kt
@@ -1,0 +1,26 @@
+package com.example.maps3dcomposedemo.widgets
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import kotlin.math.roundToInt
+
+@Composable
+fun RangeScale(range: Float, modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier
+            .background(MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.8f))
+            .padding(8.dp)
+    ) {
+        Text(
+            text = "Range: ${range.roundToInt()} m",
+            color = MaterialTheme.colorScheme.onPrimaryContainer,
+            style = MaterialTheme.typography.bodyMedium
+        )
+    }
+}
